refactor(ui): finish SurveyView split (header, floor plan, side panel, mutations)
Take SurveyView.tsx from 1649 to 702 lines, clearing the 800-line
red-flag budget. Five sibling files now share the load.

- surveyViewHelpers.tsx (190): renderSampleData + calculateMetricRange
- SurveyViewHeader.tsx (177): sticky header with start/pause/complete
  controls
- SurveyViewHeatmapSelector.tsx (192): categorised heatmap metric
  buttons (signal / interference / performance)
- SurveyViewFloorPlanPanel.tsx (369): left-column floor-plan area
  including calibration mode UI and the empty-state upload widget
- SurveyViewSidePanel.tsx (266): right-column setup checklist,
  survey settings, calibration panel, config panel, and sample list
- useSurveyMutations.ts (243): floor-plan update, config update,
  AirMapper import, settings save, and status-change callbacks

SurveyView.tsx keeps the orchestration: state, the WiFi-status and
poll-update effects, sampling logic, calibration handlers, the type
and iperf-settings setters, and the JSX render that wires the
extracted panels together. Pure relocation, no behavior change.
